Update installation docs for Svelte
I just tried to use this library after having many issues with Storybook + Svelte, but I wasn't able to have it running successfully even with an empty svelte project and following the current installation docs. 
So what I did was look for this repo examples and I see that those **examples don't follow the docs** (creating a histoire.config.js) but instead update the `vite.config.js` to include the `histoire-svelte` plugin, after I did this in my project everything works!.

I don't know if this is an old documentation or something else, but I thing the installation should work on first try so the new users don't get frustrated since they're incorrect.